[MPL] Rename ParameterProperty -> Parameter for consistency.
